Endret variabler for bedre lesbarhet

+++ CsharpProjects/Challenge-project-foreach-if-array-CSharp-main/ChallengeProject/Starter/Program.cs
@@ -66,8 +66,8 @@ foreach (string name in studentNames)
     int gradedExtraCreditAssignments = 0;
 
 
-    int sumExamScores = 0;
-    int sumExtraCreditScores = 0;
+    decimal sumExamScores = 0;
+    decimal sumExtraCreditScores = 0;
 
     decimal currentStudentGrade = 0;
     decimal currentStudentExamScore = 0;
@@ -99,13 +99,13 @@ foreach (string name in studentNames)
     }
 
     //Calculate exam score
-    currentStudentExamScore = (decimal)sumExamScores / examAssignments;
+    currentStudentExamScore = sumExamScores / examAssignments;
 
     //Calculate extra score
-    currentStudentExtraCreditScore = (decimal)sumExtraCreditScores / gradedExtraCreditAssignments;
+    currentStudentExtraCreditScore = sumExtraCreditScores / gradedExtraCreditAssignments;
 
     //Add the extra credit to the total exam score
-    currentStudentGrade = ((decimal)sumExamScores + ((decimal)sumExtraCreditScores / 10)) / examAssignments;
+    currentStudentGrade = (sumExamScores + (sumExtraCreditScores / 10)) / examAssignments;
 
 
     if (currentStudentGrade >= 97)
